Removal of refine database code from status.c and moving it respectfully to its own files
