+++ src/map/HPMmap.c
@@ -81,6 +81,8 @@
 #include "map/pet.h"
 #include "map/quest.h"
 #include "map/rodex.h"
+#include "map/refine.h"
+#include "map/refine.p.h"
 #include "map/script.h"
 #include "map/searchstore.h"
 #include "map/skill.h"
