+++ src/map/map.c
@@ -58,6 +58,7 @@
 #include "map/storage.h"
 #include "map/stylist.h"
 #include "map/rodex.h"
+#include "map/refine.h"
 #include "map/trade.h"
 #include "map/unit.h"
 #include "map/achievement.h"
@@ -6190,6 +6191,7 @@ int do_final(void)
 	atcommand->final_msg();
 	skill->final();
 	status->final();
+	refine->final();
 	unit->final();
 	bg->final();
 	duel->final();
@@ -6711,6 +6713,7 @@ int do_init(int argc, char *argv[])
 		map->read_zone_db();/* read after item and skill initialization */
 	mob->init(minimal);
 	pc->init(minimal);
+	refine->init(minimal);
 	status->init(minimal);
 	party->init(minimal);
 	guild->init(minimal);
