+++ vcproj-12/map-server.vcxproj.filters
@@ -220,6 +220,9 @@
     <ClCompile Include="..\src\map\rodex.c">
       <Filter>map</Filter>
     </ClCompile>
+    <ClCompile Include="..\src\map\refine.c">
+      <Filter>map</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\src\map\achievement.h">
@@ -499,6 +502,12 @@
       <Filter>config</Filter>
     </ClInclude>
     <ClInclude Include="..\src\map\rodex.h">
+      <Filter>map</Filter>
+    </ClInclude>
+    <ClInclude Include="..\src\map\refine.p.h">
+      <Filter>map</Filter>
+    </ClInclude>
+    <ClInclude Include="..\src\map\refine.h">
       <Filter>map</Filter>
     </ClInclude>
   </ItemGroup>
