+++ vcproj-15/map-server.vcxproj.filters
@@ -2,7 +2,7 @@
 <Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <ItemGroup>
     <ClCompile Include="..\src\map\achievement.c">
-  	  <Filter>map</Filter>
+      <Filter>map</Filter>
     </ClCompile>
     <ClCompile Include="..\src\map\atcommand.c">
       <Filter>map</Filter>
@@ -218,6 +218,9 @@
       <Filter>map</Filter>
     </ClCompile>
     <ClCompile Include="..\src\map\rodex.c">
+      <Filter>map</Filter>
+    </ClCompile>
+    <ClCompile Include="..\src\map\refine.c">
       <Filter>map</Filter>
     </ClCompile>
   </ItemGroup>
@@ -499,6 +502,12 @@
       <Filter>config</Filter>
     </ClInclude>
     <ClInclude Include="..\src\map\rodex.h">
+      <Filter>map</Filter>
+    </ClInclude>
+    <ClInclude Include="..\src\map\refine.p.h">
+      <Filter>map</Filter>
+    </ClInclude>
+    <ClInclude Include="..\src\map\refine.h">
       <Filter>map</Filter>
     </ClInclude>
   </ItemGroup>
